+++ packages/components/src/support/Story.tsx
@@ -1,11 +1,7 @@
 import React from 'react';
 import randomColor from 'randomcolor';
-import styled, { createGlobalStyle, ThemeProvider } from 'styled-components';
-import { THEME, tooltipGlobalStyles, P } from '../index';
-
-const GlobalStyle = createGlobalStyle`
-    ${tooltipGlobalStyles}
-`;
+import styled, { ThemeProvider } from 'styled-components';
+import { THEME, P } from '../index';
 
 const color = randomColor({ luminosity: 'light' });
 
@@ -22,18 +18,12 @@ const StoryWrapper = (story: any) => (
     <>
         <P size="normal">Light theme</P>
         <ThemeProvider theme={THEME.light}>
-            <Wrapper>
-                <GlobalStyle />
-                {story.children}
-            </Wrapper>
+            <Wrapper>{story.children}</Wrapper>
         </ThemeProvider>
 
         <P size="normal">Dark theme</P>
         <ThemeProvider theme={THEME.dark}>
-            <Wrapper>
-                <GlobalStyle />
-                {story.children}
-            </Wrapper>
+            <Wrapper>{story.children}</Wrapper>
         </ThemeProvider>
     </>
 );